requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
email-validator
pydantic
pyyaml
jinja2